requirement.txt
flask
python-telegram-bot
Flask-SQLAlchemy
mysql-connector-python
flask-bcrypt 
python-dotenv